Move MaintenanceGateMiddleware to core/maintenance_middleware.py
main.py was already mixing app factory, lifespan, exception handler,
router wiring, and two endpoints — adding a 44-line ASGI class with
substantive design comments next to all that made the file's "what
does it do" answer harder to skim. Move the middleware to its own
module next to core/maintenance.py (same domain), keep just the
add_middleware call in main.py, and update CLAUDE.md to point at
the new location. The middleware's invariant comments (pure-ASGI
choice + CORS-must-wrap-it ordering) travel with the class so the
load-bearing ordering note stays right next to the implementation.
